requirements.txt
dash==2.3.1
dash-mantine-components==0.10.2
dash-iconify==0.1.2
pandas==1.3.0
python-dotenv==0.20.0
gunicorn==20.0.4
databricks-sql-connector==2.4.0
pyspark==3.2.1
numpy
SQLAlchemy==1.4.44
